Use `simctl bootstatus udid -b` to boot Simulator.
`bootstatus` checks if the Simulator is booted, booting it if it is
not. This is the same behavior as before, but built into the `simctl`
command. Also, it waits until the device is booted, making the VSCode
progress message more correct (before it would say it is installing the
app during the boot process).