Fix log redirection for Simulator launch config.
It seems that `simctl launch --stdout=* ...` doesn't work with `print` in Swift.
To fix this, this change launches the app using `simctl launch --console-pty`
and pipes sdout/stderr to the files.

Unfortunately simctl buffers printing the pid (it is not printed until
the process terminates), so we need to poll for the pid. An alternative
would be to spawn using pty (using `node-pty`), or use something like
`stdbuf`/`unbuffer` to run `simctl`, but that seems more complex than
necessary.