Improve attach mode when running on device
Earlier, it was not attaching properly and launching the app in some cases.
Now we have a new method to get pid for a bundle id from ios-deploy.
Use that to get the pid and attach to that pid. Remove the earlier workaround for using launch to attach when possible.